etl/postgres_to_es/requirements.txt
pytest==6.2.5
pre-commit==2.15.0
python-dotenv==1.0.0
psycopg2-binary==2.9.5
typing-extensions==4.2.0
pydantic==1.10.7
elasticsearch==8.7.0